ASoC: tlv320aic3x: remove support for platform data
Link: https://lore.kernel.org/r/20221102232004.1721864-2-dmitry.torokhov@gmail.com
